feat(ffmpeg): merge check_ffmpeg to helpers.deps
closed #12
