+++ helpers/deps.py
@@ -1,5 +1,8 @@
+import os
 import subprocess
 import sys
+import zipfile
+
 from helpers.prompt import prompt_yes_no
 
 
@@ -69,6 +72,119 @@ def install_requirements():
     return True
 
 
+def _is_ffmpeg_installed():
+    try:
+        result = subprocess.run(['ffmpeg', '-version'], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        if result.returncode == 0:
+            print("FFmpeg is already installed.")
+            return True
+    except FileNotFoundError:
+        pass
+    return False
+
+
+def _get_latest_ffmpeg_url():
+    # Avoid import 'requests' if not needed
+    import requests
+
+    api_url = "https://api.github.com/repos/GyanD/codexffmpeg/releases/latest"
+    response = requests.get(api_url)
+    response.raise_for_status()
+    data = response.json()
+
+    for asset in data['assets']:
+        if asset['name'].endswith('full_build.zip'):
+            return asset['browser_download_url']
+    raise Exception("No FFmpeg zip file found in the latest release.")
+
+
+def _download_ffmpeg_zip(url, download_path):
+    # Avoid import 'requests' if not needed
+    import requests
+
+    response = requests.get(url)
+    response.raise_for_status()
+    with open(download_path, 'wb') as file:
+        file.write(response.content)
+    print(f"Downloaded FFmpeg ZIP to {download_path}")
+
+
+def _unzip_ffmpeg():
+    zip_path = "ffmpeg.zip"
+    extract_to = "ffmpeg"
+
+    with zipfile.ZipFile(zip_path, 'r') as zip_ref:
+        zip_ref.extractall(extract_to)
+    print(f"Extracted FFmpeg ZIP to {extract_to}")
+
+    if os.path.exists(zip_path):
+        os.remove(zip_path)
+        print(f"Deleted the ZIP file: {zip_path}")
+    else:
+        print(f"ZIP file not found: {zip_path}")
+    return extract_to
+
+
+def _find_bin_folder(extracted_folder):
+    for item in os.listdir(extracted_folder):
+        item_path = os.path.join(extracted_folder, item)
+        if os.path.isdir(item_path):
+            bin_folder = os.path.join(item_path, 'bin')
+            if os.path.exists(bin_folder):
+                return os.path.abspath(bin_folder)
+    return None
+
+
+def _find_available_powershell():
+    powershells = ['pwsh', 'powershell']
+    for ps in powershells:
+        try:
+            result = subprocess.run([ps, '-Command', 'echo test'], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            if result.returncode == 0:
+                return ps
+        except FileNotFoundError:
+            continue
+    raise Exception("Neither PowerShell nor pwsh is available on this system.")
+
+
+def _add_to_path_env(bin_folder):
+    powershell = _find_available_powershell()
+    command = f'{powershell} -Command "[Environment]::SetEnvironmentVariable(\'Path\', [Environment]::GetEnvironmentVariable(\'Path\', \'User\') + \';{bin_folder}\', \'User\')"'
+    subprocess.run(command, shell=True, check=True)
+    print(f"Updated PATH in user environment variable.")
+    print('----------------------------------------------------------------')
+    print(f"Please Close the terminal and Re open again :)")
+    print('----------------------------------------------------------------')
+    sys.exit()
+
+
+def check_ffmpeg_installed():
+    if check_requirements():
+        # Avoid import 'requests' if not needed
+        print("Please install the required packages first.")
+        install_requirements()
+
+    if not _is_ffmpeg_installed():
+        print("FFmpeg is not installed.")
+        ffmpeg_url = _get_latest_ffmpeg_url()
+
+        print(f"Latest FFmpeg URL: {ffmpeg_url}")
+        _download_ffmpeg_zip(ffmpeg_url, "ffmpeg.zip")
+        extracted_folder = _unzip_ffmpeg()
+        bin_folder = _find_bin_folder(extracted_folder)
+
+        if bin_folder:
+            print(f"Found bin folder at: {bin_folder}")
+            _add_to_path_env(bin_folder)
+
+            return True
+        else:
+            print("Failed to find the FFmpeg bin directory.")
+            return False
+    else:
+        return True
+
+
 if __name__ == "__main__":
-    missing = check_requirements()
-    _prompt_install_missing(missing)
+    install_requirements()
+    check_ffmpeg_installed()
